memide: Allocate memory for the ramdisk smartly to avoid scalability bottleneck
Allocating all the memory for the ramdisk from a single numa node hurts
scalability quite significantly. We would like to fix this by spreading the
allocations of different parts of the ramdisk across different numa nodes.

However, we can go one step further and estimate which parts of the disk will
be used heavily by which CPUs and allocate memory for those regions from the
corresponding CPUs's freelists! To achieve this, we note that we already have
per-cpu inode- and block- allocators in the filesystem. So when decompressing
the filesystem image during boot, we can distribute the inode-blocks,
bitmap-blocks and the data-blocks the same way into per-cpu buckets, and then
allocate memory for those blocks from the corresponding CPU's page-freelists.

This gives massive scalability enhancements in the lfs-largefile benchmark:

Before patch:
lfs-largefile throughput (ratio of 80 cores to 1 core) : 15x

After patch:
lfs-largefile throughput (ratio of 80 cores to 1 core) : 40x