mkfs: Encode the datablock ranges of the per-cpu journals in the superblock
For write-heavy workloads to scale in the ramdisk configuration (memide), we
need to allocate memory for the different portions of the ramdisk in a
NUMA-aware fashion. We do this for inode-blocks, bitmap-blocks and the free
data-blocks, but we can enhance this further by doing the same for the
data-blocks of the per-cpu journals.

To enable this optimization, encode the block ranges of all the per-cpu journals
in the superblock.