scalefs: Use a spinlock to protect access to journal's offset
At the moment we protect accesses to a journal's current offset by holding the
ilock for that journal's inode. However, ilock can sleep, and that would be
undesirable going forward, because we would like to access the journal's offset
while holding a spinlock.

Since updating the offset doesn't involve I/O, we can use a simple spinlock to
protect it.