scalefs: Release inodes and blocks from all transactions during group commit
Upon group commit, the current code releases the inodes and disk blocks of only
the first transaction in the batch, as opposed to all the transactions belonging
to that batch. This leads to significant pressure on inode allocations when
running benchmarks, hurting scalability. Fix it, and while at it, also take care
to deduplicate the inodes and freed blocks from the merged transactions before
releasing them, to avoid double-free bugs.