scalefs: Fix memory leak when deduplicating blocks during group commit
We recently got rid of unique-ptr for transaction diskblocks and switched to
manual memory management, but forgot to free the memory of the disk blocks that
are discarded as part of the deduplication procedure (when merging multiple
transactions). This leads to significant memory pressure during long running
benchmarks, hurting scalability. Fix the memory leak.