scalefs: Rewrite basic crash recovery using per-core journals
Every transaction committed to a per-core journal contains a start block, zero
or more dedicated address blocks, followed by one or more data blocks, and a
commit block. Upon a successful commit, the commit block's timestamp will match
the timestamp of the start block for that transaction. In any given per-core
journal, successive transactions have monotonically increasing timestamps (those
that have older timestamps are considered obsolete and hence ignored during
recovery).

Transaction dependencies can span any number of per-core journals, but they are
guaranteed to be acyclic. Dependent transactions must be applied during
crash-recovery in exactly the same order in which they were resolved and
committed to the per-core journals. So we store the time-of-commit (or
commit_tsc) in each committed transaction in the journal, to simplify resolving
dependencies during recovery. With this design, all we need to do during
recovery is read all the committed transactions from all the per-core journals,
sort them globally by their timestamps (commit_tsc) and apply them in the sorted
order. This design also helps us minimize the amount of dependency information
that must be stored in each transaction in the on-disk journal (i.e., we just
store the commit-timestamp).