scalefs: Update crash-recovery to handle journals with applied transactions
Some of the committed transactions in the per-core journals might have been
already applied by the time of the crash. Re-applying them during recovery can
corrupt the filesystem, particularly when there are dependencies between these
transactions, and some of the dependent transactions are no longer available
due to being wiped out by journal reset. An example scenario is shown below:

    Journal 0              Journal 1
 ================      ==================

 ( Txn A [Applied] )

                         Txn B [Applied]

 ( Txn C [Applied] )

                         Txn D [Committed]


Dependencies: D -> C -> B -> A  [ -> means "depends on" ]

Journal 0 was completely applied and reset (wiped) before crash; hence its
transactions are shown in parentheses, just to indicate the dependencies that
existed before the crash.

Thus, at the time of reboot, journal 0 is empty and journal 1 contains
transaction B and D. Given this state, we should skip over transaction B and
only apply transaction D, to ensure correct recovery.

So add support to note down the transactions that ought to be skipped during
recovery from each per-core journal, by adding a dedicated "skip" block at
the beginning of the journal. Every time a committed transaction is applied
from a per-core journal, its skip block is updated with the timestamp of that
transaction. This helps the recovery code skip over any transaction that was
already applied to the filesystem at the time of the crash/reboot.