scalefs: Prevent refcache from garbage-collecting newly created MemFS directories
There was a mysterious bug in ScaleFS, which used to cause a kernel crash in the
following scenario:

1. Boot sv6 with a fresh filesystem image and create a new directory under /
2. Create a large number of files under the new directory.
3. Fsync the newly created files, the new directory and / (or alternatively,
   run sync).
4. Reboot the machine using this modified filesystem image.
5. Run 'ls' on / (root directory).

The new directory would have been visible to ScaleFS towards the end of boot
(observed by instrumentation), but when running ls on /, the new directory would
unexpectedly get garbage-collected by refcache, causing a kernel panic in the
path-name resolution code. And the characteristic feature of this bug was that
this would only occur if we created a large number of files in step 2 (on the
order of hundreds of files), and if the new directory containing these files was
an immediate sub-directory of /.

It turns out that when looking up the directory entries of / (as part of
filesystem initialization), the code tries to create the ".." link (in MemFS) in
every sub-directory it finds, and as a part of this step, the as_dir() method is
invoked on the new directory. The as_dir() method in turn calls
initialize_dir(), which causes all the directory entries of *that* directory
(i.e., the new directory in this case) to be evaluated.

However, because this new directory has hundreds of files, the as_dir() method
takes slightly long to complete; this happens to be long enough to complete a
refcache epoch, and hence refcache ends up calling the onzero() method on that
directory's linkcount object, which eventually causes refcache to
garbage-collect the new directory altogether!

So when the user runs 'ls' on the root directory, the kernel's path-name
resolution code comes across a dangling directory-entry in / (in MemFS) and
panics, since the new directory had suddenly disappeared from underneath it!

Fix this by calling as_dir() on the new directory *after* we acquire an mlinkref
on that directory inside load_dir(), as opposed to doing it prematurely inside
load_dir_entry(). Also, block refcache epochs using cli (as suggested by the
mlinkref code) within load_dir_entry() to prevent potential mishaps in that
code as well.