dbench: Remove compression as well as dependency on zlib
Linking dbench with zlib on sv6 is problematic, as we already link zlib with
the kernel for decompressing the kernel image during boot.

Luckily, it turns out that the file that dbench uses to generate its load (for
the fileio backend) is actually uncompressed already (client.txt). So we don't
need compression at all to run the benchmark. So simplify dbench by getting
rid of compression and its dependency on zlib.