dbench: Make the timer thread periodic
The original dbench code uses SIGALARM and signal handlers to implement this
functionality, and it re-arms the alarm every time it fires. But when converting
this to a timer thread on sv6 (as we don't support SIGALARM), we forgot to make
it periodic, so it runs only once. Fix it.