time, hpet: Don't use the HPET for nsectime()
dbench issues a lot of calls to gettimeofday(), which boils down to nsectime(),
which in turn reads from the HPET to obtain the time. However, dbench suffers a
significant performance slowdown on sv6, and profiling with perf shows that the
majority of time is spent in hpet::read_nsec().

So, as a temporary fix, don't use the HPET for time measurements; use the ticks
instead.