dbench: Use STAT_OMIT_NLINK with stat() and fstat() in hot paths
On sv6, stat() and fstat() don't scale because, to obtain the link count of an
mnode in MemFS, we need to invoke refcache's get_consistent() method on the
mnode's link counter. And get_consistent() needs to synchronize across all cores
to obtain the true count, which makes it non-scalable.

To workaround this, sv6 offers the STAT_OMIT_NLINK flag with fstatx(), which
returns information about everything else but the link count.

The scalability of dbench suffers on sv6 due to calls to stat() and fstat() in
critical paths. So use fstatx() and statx() [the latter introduced by this
patch] with the STAT_OMIT_NLINK flag in all such paths.